[MF] Initial Sampling & Visualization
[Model Fitting] Initial Sampling and visualising the labelling by
projecting the points to the label. Refer to Carl for more detail.

+++ ModelFitting/gco-v3.0/GCoptimization.h
@@ -154,9 +154,9 @@ public:
 #ifdef GCO_ENERGYTYPE32
 	typedef int EnergyType;        // 32-bit energy total
 #else
-	typedef long long EnergyType;  // 64-bit energy total
+	typedef double EnergyType;  // 64-bit energy total
 #endif
-	typedef int EnergyTermType;    // 32-bit energy terms
+	typedef double EnergyTermType;    // 32-bit energy terms
 	typedef Energy<EnergyTermType,EnergyTermType,EnergyType> EnergyT;
 	typedef EnergyT::Var VarID;
 	typedef int LabelID;                     // Type for labels
